Fixed stuff to do with Double Precision here.
Problem was that whe double precision is used, some things are defined
twice since REAL and DOUBLE are the same. I put a #if guard around
one of these if double precision is enabled. (In single precision
this guard is not active and both single and double precisions should
be defined.)

In scalarsite_specific and scalarvecsite_specific there was the
issue of reading a complex with a param_complex_float array which
is an explicit use of float. I put #if selectors around here too,
so in double precision, param_complex_double_array is called
rather than its float friend.