Replaced initializations of SSE DP Complex temporaries of form    typedef union {         __m128d v;         double data[2];    } VD;    VD foo = { a, b };
 with __m128d foo = _mm_set_pd(b,a);

This removed some bugs in the CCMUL macro and the CCMADD macro
used in adj(M)*adj(M) multiplies. This error was picked up by the test code in qdp++/tests.
Didn't use to be an issue with GCC. In fact some of the test outside the macro still initialize the old
way.

Right now with -O2 opt level the regressions (parscalar-single-intel-double) pass except for the m_res test.