added globalness value and size checks (cirit
if this is not done I have often hard to trace
HDF5 faults. In theory the interface user could
do them, but in practive it is very often not
transparent (e.g. in a Chroma environment) what
is global and what is not and way easier to do
it here.

Similar checks might also be relevant in write routines,
I have only implemented them in the most basic cases!