fixed fatal out of memory bug, reduced memory footprint of multi[12]d writes
If multiple mpi process on the same node allocate a buffer simultaneously
(on Linux) and this exceeds the memory available, Linux will happily
allow overallocation (all processes get "valid" pointers) and the Job
will crash with untraceable segfaults.

FIX: The whole thing is unnecessary from the start since we only need
a buffer on one node (=0).

The bug is also present in specialized writes but far less likely to occur
because of the smaller (static) sizes. This is NOT fixed right now!
It is also present in LatticeObject writes, but there almost impossible
to catch. (also not fixed)