Pull #include out of QDP namespace
On GCC 6.1 (Fedora 24) qdpxx does not compile since the namespaces get
mangled by performing `#include` within the `QDP` namespace. The errors
look like this:

    In file included from /usr/include/c++/6.1.1/stdlib.h:36:0,
                     from /usr/lib/gcc/x86_64-redhat-linux/6.1.1/include/mm_malloc.h:27,
                     from /usr/lib/gcc/x86_64-redhat-linux/6.1.1/include/xmmintrin.h:34,
                     from ../include/scalarsite_sse/sse_blas_local_sumsq_double.h:15,
                     from scalarsite_sse/sse_blas_local_sumsq_double.cc:7:
    /usr/include/c++/6.1.1/cstdlib:124:11: error: '::div_t' has not been declared
       using ::div_t;
               ^~~~~
    ...
    In file included from /usr/lib/gcc/x86_64-redhat-linux/6.1.1/include/mm_malloc.h:27:0,
                     from /usr/lib/gcc/x86_64-redhat-linux/6.1.1/include/xmmintrin.h:34,
                     from ../include/scalarsite_sse/sse_blas_local_sumsq_double.h:15,
                     from scalarsite_sse/sse_blas_local_sumsq_double.cc:7:
    /usr/include/c++/6.1.1/stdlib.h:51:12: error: 'QDP::std::div_t' has not been declared
     using std::div_t;
                ^~~~~
    ...

There does not seem to be a necessity to perform the `#include` within
the namespace at all. Pulling all those out of the namespace lets it
compile on Fedora 24 with g++ 6.1.

I am not sure whether this will break the build on other machines.
Currently I do not have access to any supercomputer frontend to check.