Fixes qdpxx issue: #23 -- In chroma regressions static multi1d<LatticeXXX> objects can generate undefined behaviour at program Exit. Most visible in the case of staggered regression tests (which die with RUN_FAIL).
As @bglaessle pointed out, on destruction the multi1d<> will destroy
the Lattice Object which will call the DefaultAllocator to free memory.
However, the DefaultAllocator may be long vanished by this time.
For 2 reasons: 1) The NeverDestroy SingletonLifetime policy was changed
so it does destroy (to prevent a leak). I have reverted this to never destroy,
and those singletons are now using CreateStatic rather than CreateUsingNew.

Secondly the allocators made reference to maps, which were essentially 'global'.
I have moved the maps inside the allocator as @bglaessle suggested.

Currently results pass with the default and pool allocators (I hope, the pool
 allocator tests are still running as I write this.)