fixed gethostname issue
QDPXX provides a QDP::gethostname() which just returns the string "node".
This is supposed to have been guarded by a HAVE_GETHOSTNAME macro, but
didn't include the qdp_config.h file (it may have come in with the qdp.h file).
In any case the CMake Build didn't set this macro appropriately.
Both things are now fixed with this commit